Delete drag service and color picker files
This commit removes several files related to the drag service and color picker feature. These files were removed because they were no longer needed in the project due to changes in the requirements or design of the application. The removal of these unneeded files helps to keep the codebase clean and easier to maintain.